Applied feedback from PR
