+++ incubator/binding-pgsql/src/main/java/io/aklivity/zilla/runtime/binding/pgsql/parser/listener/SqlCreateStreamListener.java
@@ -21,14 +21,14 @@ import org.antlr.v4.runtime.TokenStream;
 
 import io.aklivity.zilla.runtime.binding.pgsql.parser.PostgreSqlParser;
 import io.aklivity.zilla.runtime.binding.pgsql.parser.PostgreSqlParserBaseListener;
-import io.aklivity.zilla.runtime.binding.pgsql.parser.module.StreamInfo;
+import io.aklivity.zilla.runtime.binding.pgsql.parser.model.StreamInfo;
 
 public class SqlCreateStreamListener extends PostgreSqlParserBaseListener
 {
-    private String name;
     private final Map<String, String> columns = new LinkedHashMap<>();
-
     private final TokenStream tokens;
+
+    private String name;
 
     public SqlCreateStreamListener(
         TokenStream tokens)
