+++ incubator/binding-pgsql/src/main/java/io/aklivity/zilla/runtime/binding/pgsql/parser/listener/SqlCreateTableTopicListener.java
@@ -23,15 +23,15 @@ import org.antlr.v4.runtime.TokenStream;
 
 import io.aklivity.zilla.runtime.binding.pgsql.parser.PostgreSqlParser;
 import io.aklivity.zilla.runtime.binding.pgsql.parser.PostgreSqlParserBaseListener;
-import io.aklivity.zilla.runtime.binding.pgsql.parser.module.TableInfo;
+import io.aklivity.zilla.runtime.binding.pgsql.parser.model.TableInfo;
 
 public class SqlCreateTableTopicListener extends PostgreSqlParserBaseListener
 {
-    private String name;
     private final Map<String, String> columns = new LinkedHashMap<>();
     private final Set<String> primaryKeys = new ObjectHashSet<>();
-
     private final TokenStream tokens;
+
+    private String name;
 
     public SqlCreateTableTopicListener(
         TokenStream tokens)
