+++ incubator/binding-pgsql/src/main/java/io/aklivity/zilla/runtime/binding/pgsql/parser/model/StreamInfo.java
@@ -12,10 +12,12 @@
  * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.aklivity.zilla.runtime.binding.pgsql.parser.module;
+package io.aklivity.zilla.runtime.binding.pgsql.parser.model;
 
-public record ViewInfo(
-        String name,
-        String select)
+import java.util.Map;
+
+public record StreamInfo(
+    String name,
+    Map<String, String> columns)
 {
 }
