+++ incubator/binding-risingwave/NOTICE
@@ -10,4 +10,5 @@ WARRANTIES OF ANY KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations under the License.
 
 This project includes:
+  zilla::incubator::binding-pgsql under Aklivity Community License Agreement
